Fix inifinite loop bug in blame.
Previously blame would get stuck in an infinite loop if it was trying
to explain a totally empty file. Conceivably this a bug in git as file
content should begin with at tab, even if it's empty.

Also adding a commit to repo1 involves fiddling lots tests so I won't
be doing that again.

Thanks to autarch for flagging this up in the first place!